backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
python-dotenv
python-jose[cryptography]
bcrypt
email-validator
sqlalchemy>=2.0
aiosqlite
pydantic-settings
openai
langchain
langchain-openai
langchain-community
langchain-chroma
langchain-huggingface
langchain-text-splitters
chromadb
tiktoken
sentence-transformers
httpx
